% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfastGSEA.R
\name{.adjustPvalues}
\alias{.adjustPvalues}
\title{Internal: Calculate adjusted p-values with multiple methods}
\usage{
.adjustPvalues(pval, method = "BH")
}
\arguments{
\item{pval}{Numeric vector of raw p-values}

\item{method}{Adjustment method: "BH", "sidak", "bonferroni", etc.}
}
\value{
Numeric vector of adjusted p-values
}
\description{
Internal: Calculate adjusted p-values with multiple methods
}
\keyword{internal}
