% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfastGSEA.R
\name{.runAdaptiveGSEA}
\alias{.runAdaptiveGSEA}
\title{Internal: Run GSEA with adaptive sampling}
\usage{
.runAdaptiveGSEA(
  pathways,
  stats,
  minSize,
  maxSize,
  scoreType,
  eps,
  gseaParam,
  pAdjustMethod = "BH"
)
}
\description{
Internal: Run GSEA with adaptive sampling
}
\keyword{internal}
