% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfastGSEA.R
\name{.runBatchGSEA}
\alias{.runBatchGSEA}
\title{Internal: Run GSEA with batch optimization}
\usage{
.runBatchGSEA(
  pathways,
  stats,
  minSize,
  maxSize,
  scoreType,
  eps,
  gseaParam,
  nAnchors,
  pAdjustMethod = "BH"
)
}
\description{
Internal: Run GSEA with batch optimization
}
\keyword{internal}
