% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{stop_health_server}
\alias{stop_health_server}
\title{Stop Health Check Server}
\usage{
stop_health_server(server)
}
\arguments{
\item{server}{Server handle returned by \code{\link{start_health_server}}}
}
\value{
\code{NULL} (invisible)
}
\description{
Stops a health check server previously started with
\code{\link{start_health_server}}.
}
\examples{
\dontrun{
srv <- start_health_server(port = 8080)
stop_health_server(srv)
}

}
